requirements.txt
pytest==8.3.4
flake8==7.1.1
pyyaml==6.0.2
python-dotenv==1.0.1
openai==1.57.4
pydantic==2.10.3
fastapi==0.115.6
mypy==1.13.0
uvicorn==0.34.0
retrying==1.3.4
numpy==2.2.0